refactor(mcpserver): split server.go into concern files (pure move)
